\name{testrvcstats-package}
\alias{testrvcstats-package}
\alias{testrvcstats}
\docType{package}
\title{
Tests rvcstats package
}
\description{
Has functions designed to test corresponding functions in rvcstats package
}
\details{
\tabular{ll}{
Package: \tab testrvcstats\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-09-14\cr
License: \tab GPLv3\cr
}
% ToDo: List of important test functions
}
\author{
Harry Ganz

Maintainer: Harry Ganz \href{hganz@rsmas.miami.edu}{mailto:hganz@rsmas.miami.edu}

}
\references{
}

\keyword{ 
package
rvc
rvcstats
testrvcstats
}
\seealso{
\link{rvcstats}
}
\examples{
%ToDo examples of test functions
}
